Removal of old types that were moved to the config module
