+++ probe-rs/src/cores/mod.rs
@@ -1,3 +1,6 @@
+use std::collections::HashMap;
+use crate::target::Core;
+
 pub mod m0;
 pub mod m33;
 pub mod m4;
@@ -17,4 +20,13 @@ impl CortexDump {
             stack,
         }
     }
+}
+
+pub fn get_core(name: impl AsRef<str>) -> Option<Box<dyn Core>> {
+    let map: HashMap<&'static str, Box<dyn Core>> = hashmap! {
+        "m0" => Box::new(self::m0::M0) as _,
+        "m4" => Box::new(self::m4::M4) as _,
+    };
+
+    map.get(&name.as_ref().to_ascii_lowercase()[..]).cloned()
 }
