+++ probe-rs/src/probe/flash/mod.rs
@@ -6,10 +6,8 @@ pub mod builder;
 pub mod download;
 pub mod flasher;
 pub mod loader;
-pub mod memory;
 
 pub use builder::*;
 pub use download::*;
 pub use flasher::*;
 pub use loader::*;
-pub use memory::*;
